Continued Work on Test_Demo
Created the TestUnit class and started to implement its functionality.
Currently, it should be able to run a function and recieve a result,
however it will not correctly parse the result nor can it take in custom
parameters yet. Additionally, work has begun an a
StringToOperationConverter class which will allow for a single expected
string to represent all possible inputs.