Started work on finishing TestUnit
The parse method has been updatd and now inlcudes a way to parse the
object given and reports when the method does not exist. The output in
TestUnit has been changed to show input vs. expected on failure and
success and the message expains which it is. All of the inverse
trigonomic functions now start with "arc" not just "a".