Util.parseConstructor now functions correctly given typed inputs for primitives.
parseConstructor now works as intended to take a string representing a
constructor and turn it into the constructor's object. Enums, booleans,
and strings to not work yet, and all primitive numbers must be typed
with a letter representing their type. For all of them, it is either the
lowercase or captial first character in their name. This typing
character goes after the number itself as if writing it in a Java
declaration.